% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{app_run}
\alias{app_run}
\title{Run the Shiny Application}
\usage{
app_run(...)
}
\arguments{
\item{...}{A series of options to be used inside the app
using \verb{shinyOptions('runtime_options')[[<option>]]}}
}
\description{
Run the Shiny Application
}
