% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{get_ch_options}
\alias{get_ch_options}
\title{Get all or one custhealth options}
\usage{
get_ch_options(which = NULL)
}
\arguments{
\item{which}{NULL (default), or the name of an option}
}
\description{
This function is to be used inside the
server and UI from your app, in order to call the
parameters passed to \code{run_app()}.
}
